refactor(modules): implement cent-based storage for monetary values
- Replace BigDecimal with Long for all monetary fields to store values in cents
- Add MoneyUtils utility class for converting between yuan and cents
- Introduce Money GraphQL scalar type for consistent money representation
- Update database schema to use amount_cents columns with bigint type
- Modify DTOs, entities, services and mappers to support cent-based storage
- Update GraphQL schema to use Money type with appropriate validation directives
- Maintain backward compatibility with amountYuan properties where needed

+++ modules/aqua-api/src/main/kotlin/dev/yidafu/aqua/api/dto/CreateOrderRequest.kt
@@ -44,7 +44,7 @@ data class CreateOrderRequest(
   @field:NotNull(message = "配送地址ID不能为空")
   val addressId: Long,
   @field:JsonProperty("amount")
-  val amount: BigDecimal,
+  val amount: Long,
   @field:JsonProperty("notes")
   @field:Size(max = 500, message = "备注长度不能超过500个字符")
   val notes: String? = null,
